+++ packages/contracts/test/plugins/governance/majority-voting/addresslist/addresslist-voting-setup.ts
@@ -5,7 +5,7 @@ import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 import {AddresslistVotingSetup} from '../../../../../typechain';
 import {deployNewDAO} from '../../../../test-utils/dao';
 import {getInterfaceID} from '../../../../test-utils/interfaces';
-import {Operation} from '../../../../core/permission/permission-manager';
+import {Operation} from '../../../../../utils/types';
 import {
   VotingSettings,
   VotingMode,
