+++ packages/contracts/test/plugins/governance/majority-voting/token/token-voting-setup.ts
@@ -5,7 +5,7 @@ import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 import {ERC20, TokenVotingSetup} from '../../../../../typechain';
 import {deployNewDAO} from '../../../../test-utils/dao';
 import {getInterfaceID} from '../../../../test-utils/interfaces';
-import {Operation} from '../../../../core/permission/permission-manager';
+import {Operation} from '../../../../../utils/types';
 import metadata from '../../../../../src/plugins/governance/majority-voting/token/build-metadata.json';
 
 import {
